layout: Add scroll area to forms and make them smaller
fixes #4

This makes the routing and accessibility map forms smaller in height and
scrollable so they can be visible on a smaller screen.

Unfortunately it does not make the whole forms responsive, as layouts in
QT designer do not seem to work properly in QGis. In order to do so, the
forms will probably have to be created programmatically instead of via a
QT designer file.